+++ Restaurant.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantHandler.cs
@@ -1,11 +1,14 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurant.Domain.Constants;
 using Restaurant.Domain.Exceptions;
+using Restaurant.Domain.Interfaces;
 using Restaurant.Domain.Repositories;
 
 namespace Restaurant.Application.Restaurants.Commands.DeleteRestaurant;
 
-public class DeleteRestaurantHandler(ILogger<DeleteRestaurantHandler> _logger, IRestaurantsRepository _restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+public class DeleteRestaurantHandler(ILogger<DeleteRestaurantHandler> _logger,
+    IRestaurantAuthorizationService restaurantAuthorizationService,IRestaurantsRepository _restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -13,6 +16,9 @@ public class DeleteRestaurantHandler(ILogger<DeleteRestaurantHandler> _logger, I
 
         var restaurant = await _restaurantsRepository.GeyRestaurantById(request.id);
         if (restaurant == null) { throw new NotFoundException($"Restaurant with that id: {request.id} does not exist  "); }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidenException();
 
         await _restaurantsRepository.DeleteRestaurant(restaurant);
